Getting the dates in the calendar for it to switch as well as separating out some of the actions and putting them in correct folders
