+++ gulp/src/analytics/components/AnalyticsApp.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {connect} from 'react-redux';
-import {doGetPageData} from '../actions/table-filter-actions';
+import {doGetPageData} from '../actions/page-loading-actions';
 import SideBar from './SideBar/SideBar';
 import Header from './Header/Header';
 import TabsContainer from './Tabs/TabsContainer';
