+++ gulp/src/sass/analytics.scss
@@ -310,6 +310,8 @@ a {
   display: none;
 }
 
+/*CSS FOR THE CALENDAR*/
+
 .calendar-panel {
   position: absolute;
   right: 200px;
@@ -319,6 +321,7 @@ a {
   background: $white;
   padding: 15px;
 }
+
 
 /*CSS FOR THE CHARTS SECTION*/
